+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 
   - Support F-Droid auto updates, refs #132
+  - Keystore directory and permissions, refs #133, #134
 
 
 ## [v20190418]